requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
Flask==2.0.2
frozenlist==1.2.0
idna==3.3
influxdb-client==1.24.0
itsdangerous==2.0.1
Jinja2==3.0.3
jsons==1.6.0
MarkupSafe==2.0.1
multidict==5.2.0
pyirsdk==1.3.2
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
Rx==3.2.0
six==1.16.0
typish==1.9.3
urllib3==1.26.7
Werkzeug==2.0.2
yarl==1.7.2
